refactor: remove inventory model and related functionality; introduce warehouseStorage model and update references accordingly
