+++ src/routes/warehouses/index.js
@@ -9,6 +9,263 @@ const router = express.Router();
 
 router.use(catchAsyncHandle(AuthMiddleware));
 
+router.get("/stock-check-requests",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get all stock check request'
+      */
+    /* #swagger.parameters['query'] = {
+        in: 'query',
+        schema: {
+            $ref: "#/components/schemas/GetAllStockCheckRequest"
+        }
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getAllStockCheckRequests)
+)
+
+router.get("/stock-check-requests/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get a stock check request'
+      */
+    /* #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check request id',
+        type: 'string'
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getStockCheckRequest)
+)
+
+router.post("/stock-check-requests",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Create a new stock check request'
+      */
+    /*  #swagger.requestBody = {
+        content: {
+            "application/json": {
+                schema: {
+                    $ref: "#/components/schemas/CreateStockCheckRequest"
+                }  
+            }
+        }
+    } 
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.createStockCheckRequest)
+)
+
+router.put("/stock-check-requests/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Update a stock check request'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check request id',
+        type: 'string'
+    }
+    #swagger.requestBody = {
+        content: {
+            "application/json": {
+                schema: {
+                    $ref: "#/components/schemas/UpdateStockCheckRequest"
+                }  
+            }
+        }
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER] }),
+    catchAsyncHandle(warehouseController.updateStockCheckRequest)
+)
+
+router.delete("/stock-check-requests/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Delete a stock check request'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check request id',
+        type: 'string'
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER] }),
+    catchAsyncHandle(warehouseController.deleteStockCheckRequest)
+)
+
+router.get("/stock-check-details",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get all stock check details'
+      */
+    /* #swagger.parameters['query'] = {
+        in: 'query',
+        schema: {
+            $ref: "#/components/schemas/GetAllStockCheckDetails"
+        }
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getAllStockCheckDetails)
+)
+
+router.get("/stock-check-details/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get a stock check detail'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check detail id',
+        type: 'string'
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getStockCheckDetail)
+)
+
+router.post("/stock-check-details",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Create a new stock check details'
+      */
+    /*  #swagger.requestBody = {
+        content: {
+            "application/json": {
+                schema: {
+                    $ref: "#/components/schemas/CreateStockCheckDetails"
+                }  
+            }
+        }
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.createStockCheckDetails)
+)
+
+router.put("/stock-check-details/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Update a stock check details'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check details id',
+        type: 'string'
+    }
+    #swagger.requestBody = {
+        content: {
+            "application/json": {
+                schema: {
+                    $ref: "#/components/schemas/UpdateStockCheckDetail"
+                }  
+            }
+        }
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.updateStockCheckDetail)
+)
+
+router.delete("/stock-check-details/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Delete a stock check details'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock check details id',
+        type: 'string'
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.deleteStockCheckDetail)
+)
+
+router.get("/storages",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get all warehouse storages'
+      */
+    /* #swagger.parameters['query'] = {
+        in: 'query',
+        schema: {
+            $ref: "#/components/schemas/GetAllInventories"
+        }
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getAllInventories)
+)
+
+router.get("/storages/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get an warehouse storage'
+      */
+    /* #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Warehouse Storage id',
+        type: 'string'
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getWarehouseStorage)
+)
+
+router.delete("/storages/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Delete an warehouse storage'
+      */
+    /*  #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Warehouse storage id',
+        type: 'string'
+    }
+*/
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER] }),
+    catchAsyncHandle(warehouseController.deleteWarehouseStorage)
+)
+
+router.get("/stock-transactions",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get all stock transactions'
+      */
+    /* #swagger.parameters['query'] = {
+        in: 'query',
+        schema: {
+            $ref: "#/components/schemas/GetAllStockTransactions"
+        }
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getAllStockTransactions)
+)
+
+router.get("/stock-transactions/:id",
+    /**
+      * #swagger.tags = ['Warehouse']
+      * #swagger.description='Get a stock transaction'
+      */
+    /* #swagger.parameters['id'] = {
+        in: 'path',
+        required: true,
+        description: 'Stock transaction id',
+        type: 'string'
+    } */
+    checkRoles({ requiredRoles: [USER_ROLES.MANAGER, USER_ROLES.INVENTORY_STAFF] }),
+    catchAsyncHandle(warehouseController.getStockTransaction)
+)
+
 router.get("/checks",
     /**
      * #swagger.tags = ['Warehouse']
