+++ swagger.js
@@ -2,10 +2,9 @@ const swaggerAutogen = require('swagger-autogen')({ openapi: '3.0.0' });
 require('dotenv').config();
 
 const { SELECT_OUTPUT, SELECT_OUTPUT_DETAILS } = require('./src/configs/output.config.js');
-const { SELECT_STOCK_DETAIL, SELECT_STOCK_REQUEST, SELECT_STOCK_TRANSACTION, SELECT_INVENTORY } = require('./src/configs/inventory.config.js');
 const { SELECT_ITEM } = require('./src/configs/item.config.js');
 const { SELECT_USER } = require('./src/configs/user.config.js');
-const { SELECT_WAREHOUSE, SELECT_WAREHOUSE_CHECK, SELECT_WAREHOUSE_CHECK_DETAIL } = require('./src/configs/warehouse.config.js');
+const { SELECT_WAREHOUSE, SELECT_WAREHOUSE_CHECK, SELECT_STOCK_DETAIL, SELECT_STOCK_REQUEST, SELECT_STOCK_TRANSACTION, SELECT_WAREHOUSE_STORAGE } = require('./src/configs/warehouse.config.js');
 
 const doc = {
     host: `localhost:${process.env.DEV_APP_PORT}`,            // by default: 'localhost:3000'
@@ -32,10 +31,6 @@ const doc = {
         },
         {
             name: 'Warehouse',
-            description: ''
-        },
-        {
-            name: 'Inventory',
             description: ''
         },
         {
@@ -140,14 +135,14 @@ const doc = {
                 condition: "Good",
                 status: 'Done'
             },
-            GetAllInventories: {
+            GetAllWarehouseStorages: {
                 limit: 10,
                 sort: 'ctime',
                 page: 1,
                 filter: {
                     isDeleted: false
                 },
-                select: SELECT_INVENTORY,
+                select: SELECT_WAREHOUSE_STORAGE,
                 expand: 'warehouse item'
             },
             GetAllStockTransactions: {
@@ -233,7 +228,7 @@ const doc = {
                     isDeleted: false
                 },
                 select: SELECT_OUTPUT,
-                expand: 'customer warehouse reportStaff manager inventoryStaff outputDetails'
+                expand: 'customer reportStaff manager inventoryStaff'
             },
             GetAllOutputDetails: {
                 limit: 10,
